cambios pelucheee version ultimate

+++ src/app/config-empresa/config-empresa.component.ts
@@ -3,6 +3,7 @@ import { Empresa } from '../modelo/empresa';
 import { EmpresaService } from '../_services/empresa.service';
 import { Router } from '@angular/router';
 import { AllScriptServiceService } from '../all-script-service.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-config-empresa',
@@ -38,13 +39,28 @@ export class ConfigEmpresaComponent implements OnInit {
 
 
 
-  eliminarEmpresa(idConfig:number){
-    this.empresaServicio.eliminarEmpresa(idConfig).subscribe(dato =>{
-      console.log(dato);
-      this.obtenerEmpresa();
-    });
-  }
 
+
+  eliminarEmpresa(idConfig:number){
+    Swal.fire({
+      title: '¿Estás seguro de que deseas eliminar?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Eliminar',
+      cancelButtonText: 'Cancelar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.empresaServicio.eliminarEmpresa(idConfig).subscribe(dato =>{
+          console.log(dato);
+          this.obtenerEmpresa();
+        });
+        Swal.fire('Eliminado!', 'El elemento ha sido eliminado.', 'success');
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        // Acción cuando se hace clic en "Cancelar" o se cierra el SweetAlert
+        Swal.fire('Cancelado', 'La eliminación ha sido cancelada.', 'info');
+      }
+    });
+    }
   actualizarEmpresa(id:number){
     this.router.navigate(['actualizar-empresa',id]);
   }
